+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+
+### Added
+- Parallel prime finding methods and a "multicore" feature ([#60])
+
+[#60]: https://github.com/entropyxyz/crypto-primes/pull/60
+
 ## [0.6.0-pre.2] - 2024-10-18
 
 ### Changed